requirements.txt
cycler==0.10.0
descartes==1.1.0
kiwisolver==1.2.0
matplotlib==3.2.1
mizani==0.6.0
numpy==1.18.4
palettable==3.3.0
pandas==1.0.3
patsy==0.5.1
Pillow==7.1.2
plotnine==0.6.0
progressbar2==3.51.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
scipy==1.4.1
six==1.14.0
statsmodels==0.11.1
xlrd==1.2.0
